requirements.txt
asgiref==3.6.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
Flask==2.3.2
fonttools==4.39.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.3
oauth2-provider==0.0
packaging==23.1
Pillow==9.5.0
prettytable==3.7.0
psycopg2==2.9.6
pyotp==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.3
tk==0.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
wcwidth==0.2.6
Werkzeug==2.3.3
whitenoise==6.4.0
